azure-functions/finetuning-pipeline/requirements.txt
azure-functions
azure-storage-blob
azure-identity
openai
typing-extensions
